ollama-stream-chat/requirements.txt
psycopg2_binary==2.9.10
sentence_transformers==4.1.0
streamlit==1.45.1
